Backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.30.1
python-dotenv==1.0.1
pydantic[email]==2.11.1
aiosmtplib==3.0.1
httpx==0.27.0